refactor: move av components to their own folders
