+++ src/ui/cards/AvCard/index.ts
@@ -1,0 +1,1 @@
+export { default as AvCard, type AvCardProps } from './AvCard.vue'
