+++ src/ui/cards/index.ts
@@ -1,1 +1,1 @@
-export { default as AvCard } from './AvCard.vue'
+export * from './AvCard'
