--- src/ui/cards/types.ts
@@ -1,4 +1,0 @@
-export interface AvCardProps {
-  borderColor?: string
-  titleBackground?: string
-}
