+++ src/ui/header/AvHeader/index.ts
@@ -1,0 +1,1 @@
+export { default as AvHeader, type AvHeaderProps } from './AvHeader.vue'
