+++ src/ui/interaction/buttons/AvButton/AvButton.vue
@@ -1,7 +1,12 @@
 <script setup lang="ts">
-import type { VIcon } from '@gouvminint/vue-dsfr'
-import type { AvButtonProps } from './types'
+import type { DsfrButtonProps, VIcon } from '@gouvminint/vue-dsfr'
 import { MDI_ICONS } from '@/ui'
+
+export type AvButtonProps = {
+  variant?: 'primary' | 'secondary' | 'tertiary' | 'tertiary-no-outline'
+  isLoading?: boolean
+  color?: string
+} & Pick<DsfrButtonProps, 'label' | 'disabled' | 'size' | 'icon' | 'iconRight' | 'iconOnly' | 'onClick'>
 
 const props = withDefaults(defineProps<AvButtonProps>(), {
   variant: 'primary',
