+++ src/ui/interaction/buttons/AvButton/index.ts
@@ -1,0 +1,1 @@
+export { default as AvButton, type AvButtonProps } from './AvButton.vue'
