+++ src/ui/interaction/buttons/index.ts
@@ -1,1 +1,1 @@
-export { default as AvButton } from './AvButton.vue'
+export * from './AvButton'
