+++ src/ui/navigation/AvNavigation/index.ts
@@ -1,0 +1,1 @@
+export { default as AvNavigation, type AvNavigationProps } from './AvNavigation.vue'
