--- src/ui/navigation/types.ts
@@ -1,5 +1,0 @@
-import type { DsfrNavigationProps } from '@gouvminint/vue-dsfr'
-
-export interface AvNavigationProps {
-  navItems: DsfrNavigationProps['navItems']
-}
